Move addPayment to BatchStrategy.
